# backend/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"UserRole": {
				"type": "string",
				"enum": [
					"recruiter",
					"admin",
					"applicant"
				]
			},
			"NotificationPreferences": {
				"properties": {
					"screening": {
						"type": "boolean"
					},
					"applicants": {
						"type": "boolean"
					},
					"export": {
						"type": "boolean"
					},
					"system": {
						"type": "boolean"
					}
				},
				"required": [
					"screening",
					"applicants",
					"export",
					"system"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ThemePreference": {
				"type": "string",
				"enum": [
					"light",
					"dark",
					"system"
				]
			},
			"LanguagePreference": {
				"type": "string",
				"enum": [
					"en",
					"fr",
					"rw"
				]
			},
			"IUserResponse": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"googleId": {
						"type": "string"
					},
					"isEmailVerified": {
						"type": "boolean"
					},
					"profilePicture": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"notificationPreferences": {
						"$ref": "#/components/schemas/NotificationPreferences"
					},
					"themePreference": {
						"$ref": "#/components/schemas/ThemePreference"
					},
					"languagePreference": {
						"$ref": "#/components/schemas/LanguagePreference"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"email",
					"firstName",
					"lastName",
					"role",
					"isEmailVerified",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthResponse": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/IUserResponse"
					},
					"token": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"user",
					"token",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ProfileResponse": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/IUserResponse"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"user",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UpdateProfileRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"profilePicture": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"RefreshTokenRequest": {
				"properties": {
					"userId": {
						"type": "string"
					}
				},
				"required": [
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GoogleAuthRequest": {
				"properties": {
					"idToken": {
						"type": "string"
					}
				},
				"required": [
					"idToken"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "backend-api",
		"version": "1.0.0",
		"description": "Simple backend API with Hello World endpoint using TypeScript, tsoa, and Mongoose",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Bikorimana Saveur"
		}
	},
	"paths": {
...
		"/auth/me": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfileResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"patch": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProfileResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProfileRequest"
							}
						}
					}
				}
			}
		},
...
		"/auth/refresh": {
			"post": {
				"operationId": "RefreshToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RefreshTokenRequest"
							}
						}
					}
				}
			}
		},
		"/auth/google": {
			"post": {
				"operationId": "GoogleAuth",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GoogleAuthRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}